update 1.6
What has changed:
Import Pixmap:

We added from Components.Pixmap import Pixmap to use the appropriate widget to display the image.
Checking the existence of the file:

In set_logo we check whether the logo file exists before we set it.
Using setPixmapFromFile:

This enables the image from the file to be displayed on the Pixmap widget.
Transparency and appearance:

We kept the options transparent="1" and alphatest="on" in the skin definition.